feat: Enhance tracing mechanism with improved start and end time handling and performance

+++ src/main/java/com/cmdev/profiler/instrument/TimerTracingUtils.java
@@ -25,7 +25,7 @@ public class TimerTracingUtils {
 
         if (threadIdLocal != null) {
             int[] depthHolder = deepOfTheMessage.get();
-            long depthValue;
+            int depthValue;
             if (!traceInfos.isEnd()) {
                 depthValue = depthHolder[0]++;
             } else {
